+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Introduce error estimation in scorer for all metrics
 * Fix validation score delta that was applied in the incorrect direction
 * Fix LuaJIT out of memory issues in `learn_bpe.lua`
+* More graceful error handling
 
 ## [v0.8.0](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.8.0) (2017-06-28)